requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
boto==2.48.0
certifi==2017.7.27.1      # via requests
chardet==3.0.4            # via requests
django-recaptcha==1.3.1
django-registration-redux==1.8
django-ses==0.8.3.1
django-suit==0.2.25
django==1.11.7
idna==2.6                 # via requests
mysqlclient==1.3.12
pytz==2017.3              # via django, django-ses
raven==6.3.0
requests==2.18.4
urllib3==1.22             # via requests
